게시물수정,삭제시권한확인

+++ src/com/jtp/example/mysqlTextBoard/controller/ArticleController.java
@@ -74,7 +74,7 @@ public class ArticleController extends Controller {
 	}
 
 	private void doDelete(String cmd) {
-		System.out.println("== 게시물 삭 ==");
+		System.out.println("== 게시물 삭제 ==");
 		if (Container.session.isLogined() == false) {
 			System.out.println("로그인 후 이용해주세요.");
 			return;
@@ -89,13 +89,18 @@ public class ArticleController extends Controller {
 			System.out.println("존재하지 않는 게시물 입니다.");
 			return;
 		}
+		if (article.memberId != Container.session.getLoginedMemberId()) {
+			System.out.println("권한이 없습니다.");
+			return;
+		}
+
 		articleService.delete(input);
 		System.out.printf("%d번 게시물을 삭제하였습니다.\n", input);
 
 	}
 
 	private void doWrite(String cmd) {
-		System.out.println("== 게시물 작 ==");
+		System.out.println("== 게시물 작성 ==");
 		if (Container.session.isLogined() == false) {
 			System.out.println("로그인 후 이용해주세요.");
 			return;
@@ -125,16 +130,23 @@ public class ArticleController extends Controller {
 			System.out.println("로그인 후 이용해주세요.");
 			return;
 		}
+		
 		if (cmd.split("article")[1].equals(" modify")) {
 			return;
 		}
 		int input = Integer.parseInt(cmd.split(" ")[2]);
 
 		Article article = articleService.getArticle(input);
+		
 		if (article == null) {
 			System.out.println("존재하지 않는 게시물 입니다.");
 			return;
 		}
+		if (article.memberId != Container.session.getLoginedMemberId()) {
+			System.out.println("권한이 없습니다.");
+			return;
+		}
+
 		Member member = memberService.getMemberById(article.memberId);
 		String writer = member.name;
